ApiUser/requirements.txt
Flask==1.0.2
pyzmq==18.0.1
jsonschema==3.0.1
pymongo==3.7.2
Flask-PyMongo==2.2.0
flasgger==0.9.2